modularizando o componente principal
