+++ ta-gui/src/app/aluno.service.ts
@@ -20,12 +20,20 @@ export class AlunoService {
      return !this.alunos.find(a => a.cpf == cpf);
   }
 
-  atualizar(aluno:Aluno): void {
+  atualizar(aluno: Aluno): void {
     aluno = aluno.clone();
     for (let a of this.alunos) {
         if (a.cpf == aluno.cpf) {
            a.metas = aluno.metas;
         }
     }
+  }
+
+  getAlunos(): Aluno[] {
+    var result: Aluno[] = [];
+    for (let a of this.alunos) {
+      result.push(a.clone());
+    }
+    return result;
   }
 }
