+++ ta-gui/src/app/alunos.component.html
@@ -3,13 +3,13 @@
   <table>
  <tr>
     <td><label>Nome </label></td>
-    <td><input [(ngModel)]="aluno.nome" name="namebox"></td>
+    <td><input [(ngModel)]="aluno.nome" ></td>
     <td><button (click)="criarAluno(aluno)">Adicionar</button></td>
  </tr>
  <tr>
     <td><label>CPF </label></td>
-    <td><input [(ngModel)]="aluno.cpf" name="cpfbox"></td>
-    <td name="msgcpfexistente" *ngIf="cpfduplicado">
+    <td><input [(ngModel)]="aluno.cpf" ></td>
+    <td *ngIf="cpfduplicado">
       Já existe um aluno com esse CPF
     </td>
  </tr>
@@ -25,9 +25,9 @@
     <th>CPF</th>
     <th>e-mail</th>
  </tr>
- <tr *ngFor="let a of alunos" name="alunolist">
-    <td name="nomelist">{{a.nome}}</td>
-    <td name="cpflist">{{a.cpf}}</td>
+ <tr *ngFor="let a of alunos">
+    <td>{{a.nome}}</td>
+    <td>{{a.cpf}}</td>
     <td>{{a.email}}</td>
  </tr>
  </table>
