+++ ta-gui/src/app/alunos.component.ts
@@ -1,6 +1,8 @@
 import { Component, OnInit } from '@angular/core';
-import { Aluno } from '../../../common/aluno';
-import { AlunoService } from './aluno.service';
+ import { NgModule } from '@angular/core';
+
+  import { Aluno } from './aluno';
+ import { AlunoService } from './aluno.service';
 
   @Component({
    selector: 'app-root',
@@ -8,38 +10,27 @@ import { AlunoService } from './aluno.service';
    styleUrls: ['./alunos.component.css']
  })
  export class AlunosComponent implements OnInit {
-
-    aluno: Aluno = new Aluno();
-    alunos: Aluno[] = [];
-    cpfduplicado: boolean = false;
-
     constructor(private alunoService: AlunoService) {}
 
-     criarAluno(a: Aluno): void {
-       this.alunoService.criar(a)
-              .subscribe(
-                ar => {
-                  if (ar) {
-                    this.alunos.push(ar);
-                    this.aluno = new Aluno();
-                  } else {
-                    this.cpfduplicado = true;
-                  } 
-                },
-                msg => { alert(msg.message); }
-              );
-    } 
+    aluno: Aluno = new Aluno();
+    alunos: Aluno[];
+    cpfduplicado: boolean = false;
 
-    onMove(): void {
+     criarAluno(a: Aluno): void {
+      if (this.alunoService.criar(a)) {
+        this.alunos.push(a);
+        this.aluno = new Aluno();
+      } else {
+        this.cpfduplicado = true;
+      }
+    }
+
+     onMove(): void {
        this.cpfduplicado = false;
     }
 
      ngOnInit(): void {
-       this.alunoService.getAlunos()
-             .subscribe(
-               as => { this.alunos = as; },
-               msg => { alert(msg.message); }
-              );
-     }
+      this.alunos = this.alunoService.getAlunos();
+    }
 
   }
