+++ ta-gui/src/app/app.component.html
@@ -1,46 +1,14 @@
-<html (mousemove)="onMove()">
+<html>
+
 <h1>
   Teaching Assistant
 </h1>
 
-<h2>
-  Alunos
-</h2>
+<nav>
+    <a routerLink="/alunos">Alunos</a>
+    <a routerLink="/metas">Metas</a>
+</nav>
 
-<table>
-<tr>
-   <td><label>Nome </label></td>
-   <td><input [(ngModel)]="aluno.nome" ></td>
-   <td><button (click)="criarAluno(aluno)">Adicionar</button></td>
-</tr>
-<tr>
-   <td><label>CPF </label></td>
-   <td><input [(ngModel)]="aluno.cpf" ></td>
-   <td *ngIf="cpfduplicado">
-     Já existe um aluno com esse CPF
-   </td>
-</tr>
-<tr>
-   <td><label>e-mail </label></td>
-   <td><input [(ngModel)]="aluno.email" ></td>
-</tr>
-</table>
-
-<table class="alunos">
-<tr *ngIf="alunos.length > 0">
-   <th>Nome</th>
-   <th>CPF</th>
-   <th>e-mail</th>
-   <th>Requisitos</th>
-   <th>Gerência de Configuração</th>
-</tr>
-<tr *ngFor="let a of alunos">
-   <td>{{a.nome}}</td>
-   <td>{{a.cpf}}</td>
-   <td>{{a.email}}</td>
-   <td><input [(ngModel)]="a.metas['requisitos']" (change)="atualizarAluno(a)"></td>
-   <td><input [(ngModel)]="a.metas['gerDeConfiguracao']" (change)="atualizarAluno(a)"></td>
-</tr>
-</table>
+<router-outlet></router-outlet>
 
 </html>
