+++ ta-gui/src/app/app.component.ts
@@ -29,8 +29,4 @@ export class AppComponent {
       this.cpfduplicado = false;
    }
 
-   atualizarAluno(aluno: Aluno): void {
-      this.alunoService.atualizar(aluno);
-   }
-
 }
