+++ ta-gui/src/app/app.module.ts
@@ -1,19 +1,34 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
 import { FormsModule } from '@angular/forms';
+import { RouterModule }   from '@angular/router';
 
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
+import { MetasComponent } from './metas.component';
+import { AlunosComponent } from './alunos.component';
 import { AlunoService } from './aluno.service';
 
 @NgModule({
   declarations: [
     AppComponent,
+    MetasComponent,
+    AlunosComponent
   ],
   imports: [
     BrowserModule,
     AppRoutingModule,
-    FormsModule
+    FormsModule,
+    RouterModule.forRoot([
+      {
+        path: 'metas',
+        component: MetasComponent
+      },
+      {
+        path: 'alunos',
+        component: AlunosComponent
+      }
+    ])
   ],
   providers: [AlunoService],
   bootstrap: [AppComponent]
